requirements.txt
python-dateutil==2.8.1
python-Levenshtein==0.12.0
tranco==0.5
pyenchant==3.1.1
tldextract==2.2.2